diagho_project/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
exceptiongroup==1.2.2
idna==3.7
iniconfig==2.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
